ai-backend/requirements.txt
accelerate==1.13.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
comm==0.2.3
confection==1.3.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.3.1
distro==1.9.0
dotenv==0.9.9
emoji==2.15.0
executing==2.2.1
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-vision==3.12.1
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.4.2
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
installer==0.7.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
langchain==1.2.13
langchain-core==1.2.22
langchain-openai==1.1.12
langdetect==1.0.9
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
murmurhash==1.0.15
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
olefile==0.47
onnx==1.20.1
onnxruntime==1.24.4
openai==2.29.0
opencv-python==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
parso==0.8.6
pdf2image==1.17.0
pdfminer.six==20260107
pexpect==4.9.0
pi_heif==1.3.0
pikepdf==10.5.1
pillow==12.1.1
platformdirs==4.9.4
preshed==3.0.13
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.9.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
timm==1.0.26
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.21.5
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
wcwidth==0.6.0
weasel==1.0.0
webencodings==0.5.1
wrapt==1.17.3
xlsxwriter==3.2.9
xxhash==3.6.0
zstandard==0.25.0
